line, and those below, will be ignored--
M    bibliotecas/eVector/primitive_types.h
M    C++/qImageTransform/qimageshow.cpp
M    C++/qImageTransform/pixelManipulation_qt.h
M    C++/qImageTransform/qImageTransform.pro
M    C++/qImageTransform/ui_controlpainel_qt.h
M    C++/curveFitting/cubiccurvefitter.cpp
M    C++/curveFitting/cubicsegment.cpp
M    C++/curveFitting/cubicsegment.h
M    C++/curveFitting/drawarea.cpp

+++ bibliotecas/eVector/primitive_types.h
@@ -18,6 +18,6 @@
   typedef double  fp64;
 
   typedef sint32   word;
-  typedef fp64     real;
+  typedef fp32     real;
 
 #endif  // __PRIMITIVE_TYPES_HPP__
